+++ src/ecies/curve_25519/ecies_x25519_xchacha20.rs
@@ -297,8 +297,6 @@ mod tests {
             Some(&b"Corrupted authenticated data"[..]),
         );
 
-        println!("{not_decrypted:?}");
-
         assert!(matches!(
             not_decrypted,
             Err(CryptoCoreError::DecryptionError)